requirements.txt
fluxer.py
python-dotenv
requests
colorama